refactored currency and currencies class

+++ src/de/wagnst/tpe/exercises/exercise2/master/Currencies.java
@@ -1,8 +1,8 @@
 package de.wagnst.tpe.exercises.exercise2.master;
 
 /**
- * contains defined objects of currency with its exchange rate to dollar,
- * name, currency symbol and an information if it has a subunit 
+ * Contains defined objects of currency with its exchange rate to dollar,
+ * name, currency symbol and an information if it has a subunit
  *
  * @author maxGross, wagnst
  */
@@ -11,12 +11,11 @@ public final class Currencies {
     /**
      * Definition of exchange rates to basis dollar
      */
-    private static final int EXCHANGE_DOLLAR = 10000;
-    private static final int EXCHANGE_EURO = 12690;
-    private static final int EXCHANGE_YEN = 91;
-    private static final int EXCHANGE_RUBEL = 255;
-    private static final int EXCHANGE_CHF = 10509;
-    
+    private static final long EXCHANGE_DOLLAR = 10000;
+    private static final long EXCHANGE_EURO = 12690;
+    private static final long EXCHANGE_YEN = 91;
+    private static final long EXCHANGE_RUBEL = 255;
+    private static final long EXCHANGE_CHF = 10509;
 
     /**
      * US-Dollar
@@ -30,19 +29,19 @@ public final class Currencies {
             EXCHANGE_EURO, true);
 
     /**
-     * Japanischer Yen
+     * Japanese Yen
      */
     public static final Currency YEN = new Currency("Yen", "¥", EXCHANGE_YEN,
             false);
 
     /**
-     * Rubel der russischen Föderation
+     * Rubel of the Russian Federation
      */
     public static final Currency RUBEL = new Currency("Rubel", "RUB",
             EXCHANGE_RUBEL, false);
 
     /**
-     * Schweizer Franken
+     * Swiss Frank
      */
     public static final Currency CHF = new Currency("Schweizer Franken", "CHF",
             EXCHANGE_CHF, true);
